pyright gh actions
