+++ pyproject.toml
@@ -29,6 +29,7 @@ dnspython = "^2.1.0"
 mkdocs-material = "^7.1.0"
 pydoc-markdown = "^3.13.0"
 flake8 = "^3.9.1"
+pyright = "^0.0.2"
 
 
 [build-system]
@@ -77,3 +78,6 @@ ignore_missing_imports = true
 [[tool.mypy.overrides]]
 module = "toml"
 ignore_missing_imports = true
+
+[tool.pyright]
+include = ["tests/typing"]
